pyproject.toml
[tool.poetry]
name = "api-test"
version = "0.1.0"
description = ""
authors = ["Aleksei Chistov <aleksey.chistov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"
celery = {extras = ["redis"], version = "^5.5.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90

[tool.isort]
profile = "black"
